requirement.txt
numpy==2.2.6
opencv-python==4.13.0.92
pillow==12.1.1
scikit-learn==1.7.2
seaborn==0.13.2